--- src/environments/api-keys.env
@@ -1,6 +1,0 @@
-CTP_PROJECT_KEY=bon747jour
-CTP_CLIENT_SECRET=FdHmxPyT12o4qb0VWg4z5k_glAZI5vbt
-CTP_CLIENT_ID=SiyGK3A8uQpMq5CgHb3egU9K
-CTP_AUTH_URL=https://auth.europe-west1.gcp.commercetools.com
-CTP_API_URL=https://api.europe-west1.gcp.commercetools.com
-CTP_SCOPES=manage_project:bon747jour
