+++ src/pages/SignInPage/SignInPage.tsx
@@ -1,6 +1,6 @@
 import { Button } from 'primereact/button';
 import { useNavigate } from 'react-router-dom';
-import { FormSingIn } from '../../components/Forms/SignInForm';
+import { SignInForm } from '../../components/Forms/SignInForm';
 import { PAGES } from '../../constants/pages';
 import styles from './SignInPage.module.scss';
 
@@ -9,7 +9,7 @@ export const SignInPage = (): JSX.Element => {
 
   return (
     <div className={styles.page}>
-      <FormSingIn />
+      <SignInForm />
       <h4 className="center mb-2 pl-2 pr-2 text-center">
         If you are not registered, please register in our store.
       </h4>
